+++ examples/test-server-as-promised-ts.ts
@@ -2,11 +2,10 @@
 
 import fs from 'fs'
 import net from 'net'
-import NullWritable from 'null-writable'
 import PromiseReadable from 'promise-readable'
 import { Readable } from 'stream'
 
-import { SMTPServerAddress, SMTPServerAsPromised, SMTPServerAsPromisedOptions, SMTPServerSession } from '../src/smtp-server-as-promised'
+import { SMTPServerAddress, SMTPServerAsPromised, SMTPServerAsPromisedOptions, SMTPServerAuthentication, SMTPServerAuthenticationResponse, SMTPServerSession } from '../src/smtp-server-as-promised'
 
 interface ArgvOptions {
   [key: string]: string
@@ -16,71 +15,69 @@ interface Session extends SMTPServerSession {
   messageLength?: number
 }
 
-async function onConnect (session: Session): Promise<void> {
-  console.info(`[${session.id}] onConnect`)
+interface MySMTPServerOptions extends SMTPServerAsPromisedOptions {
+  password?: string
 }
 
-// async function onAuth (auth: SMTPServerAsPromised.Authentication, session: Session): Promise<SMTPServerAsPromised.AuthenticationResponse> {
-//   if (auth.method === 'PLAIN' && auth.username === 'username' && auth.password === 'password') {
-//     return { user: auth.username }
-//   } else {
-//     throw new Error('Invalid username or password')
-//   }
-// }
+class MySMTPServer extends SMTPServerAsPromised {
+  password?: string
 
-async function onMailFrom (from: SMTPServerAddress, session: Session): Promise<void> {
-  const tlsDebug = session.tlsOptions ? JSON.stringify(session.tlsOptions) : ''
-  console.info(`[${session.id}] onMailFrom ${from.address} ${session.openingCommand} ${session.transmissionType} ${tlsDebug}`)
-  if (from.address.split('@')[1] === 'spammer.com') {
-    // code 421 disconnects SMTP session immediately
-    throw Object.assign(new Error('we do not like spam!'), { responseCode: 421 })
-  } else if (from.address.split('@')[0] === 'bounce') {
-    throw Object.assign(new Error('fatal'), { responseCode: 500 })
+  constructor (options: MySMTPServerOptions) {
+    super(options)
+    this.password = options.password
   }
-}
 
-async function onRcptTo (to: SMTPServerAddress, session: Session): Promise<void> {
-  console.info(`[${session.id}] onRcptTo ${to.address}`)
-}
+  protected async onConnect (session: Session): Promise<void> {
+    console.info(`[${session.id}] onConnect`)
+  }
 
-async function onData (stream: Readable, session: Session): Promise<void> {
-  console.info(`[${session.id}] onData started`)
+  protected async onAuth (auth: SMTPServerAuthentication, session: Session): Promise<SMTPServerAuthenticationResponse> {
+    console.info(`[${session.id} onAuth ${auth.method} ${auth.username} ${auth.password}`)
+    if (auth.method === 'PLAIN' && auth.username === 'username' && auth.password === this.password) {
+      return { user: auth.username }
+    } else {
+      throw new Error('Invalid username or password')
+    }
+  }
 
-  try {
+  protected async onMailFrom (from: SMTPServerAddress, session: Session): Promise<void> {
+    const tlsDebug = session.tlsOptions ? JSON.stringify(session.tlsOptions) : ''
+    console.info(`[${session.id}] onMailFrom ${from.address} ${session.openingCommand} ${session.transmissionType} ${tlsDebug}`)
+    if (from.address.split('@')[1] === 'spammer.com') {
+      // code 421 disconnects SMTP session immediately
+      throw Object.assign(new Error('we do not like spam!'), { responseCode: 421 })
+    } else if (from.address.split('@')[0] === 'bounce') {
+      throw Object.assign(new Error('fatal'), { responseCode: 500 })
+    }
+  }
+
+  protected async onRcptTo (to: SMTPServerAddress, session: Session): Promise<void> {
+    console.info(`[${session.id}] onRcptTo ${to.address}`)
+  }
+
+  protected async onData (stream: Readable, session: Session): Promise<void> {
+    console.info(`[${session.id}] onData started`)
     const promiseStream = new PromiseReadable(stream)
     const message = await promiseStream.readAll()
     console.info(`[${session.id}] onData read\n${message}`)
 
     session.messageLength = message ? message.length : 0
     console.info(`[${session.id}] onData finished after reading ${session.messageLength} bytes`)
-  } catch (e) {
-    // read it to the end
-    stream.pipe(new NullWritable())
-    // rethrow original error
-    throw e
   }
-}
 
-async function onClose (session: Session): Promise<void> {
-  console.info(`[${session.id}] onClose`)
-}
+  protected async onClose (session: Session): Promise<void> {
+    console.info(`[${session.id}] onClose`)
+  }
 
-async function onError (err: Error): Promise<void> {
-  console.info('Server error:', err)
+  protected async onError (err: Error): Promise<void> {
+    console.info('Server error:', err)
+  }
 }
 
 async function main (): Promise<void> {
   // Usage: node server.js opt1=value1 opt2=value2...
-  const defaultOptions: SMTPServerAsPromisedOptions & net.ListenOptions = {
-    disabledCommands: ['AUTH'],
+  const defaultOptions: MySMTPServerOptions & net.ListenOptions = {
     hideSTARTTLS: true,
-    // onAuth,
-    onClose,
-    onConnect,
-    onData,
-    onError,
-    onMailFrom,
-    onRcptTo,
     port: 2525
   }
 
@@ -88,11 +85,15 @@ async function main (): Promise<void> {
 
   const options = { ...defaultOptions, ...userOptions } as SMTPServerAsPromisedOptions & net.ListenOptions
 
+  if (!userOptions.password) {
+    options.disabledCommands = ['AUTH']
+  }
+
   options.ca = typeof options.ca === 'string' ? fs.readFileSync(options.ca) : undefined
   options.cert = typeof options.cert === 'string' ? fs.readFileSync(options.cert) : undefined
   options.key = typeof options.key === 'string' ? fs.readFileSync(options.key) : undefined
 
-  const server = new SMTPServerAsPromised(options)
+  const server = new MySMTPServer(options)
   const address = await server.listen(options)
   console.info(`Listening on [${address.address}]:${address.port}`)
 }